feat: Add slash command
